move scenarios CR-T148 and CR-T149 over to the testCaseEndpoints feature as they are testing getCaseByUprn
