+++ src/test/resources/integrationtests/address/testAddressEndpoints.feature
@@ -48,27 +48,15 @@ Feature: Test Contact centre Address Endpoints
       | "Boaty McBoat Face " |
       | "Strawberry Laces"   |
 
-  Scenario: AIMS endpoint to provide region code, address type and establishment type
+  Scenario: [CR-T147] AIMS endpoint to provide region code, address type and establishment type
     Given the respondent calls the CC with a fulfilment request
     And the respondent address exists in AIMS
     When the CC agent searches for the address
     Then the CC SVC returns address attributes with region code, address type and establishment type
 
-  Scenario: Publish a new address event to RM
-    Given the CC agent has confirmed the respondent address
-    And the case service does not have any case created for the address in question
-    When Get/Case API returns a 404 error because there is no case found
-    And CC SVC creates a fake Case with the address details from AIMS
-    Then the CC SVC must publish a new address event to RM with the fake CaseID
-    
-  #Scenario: Publish a fulfilment request event to RM when case does not exist in RM but in firestore cache
+  #Scenario: Publish a new address event to RM
     #Given the CC agent has confirmed the respondent address
     #And the case service does not have any case created for the address in question
-    #When Get/Case API returns a 404 error because the case was not found
-    #And CC SVC creates a fake case with the address details from AIMS
+    #When Get/Case API returns a 404 error because there is no case found
+    #And CC SVC creates a fake Case with the address details from AIMS
     #Then the CC SVC must publish a new address event to RM with the fake CaseID
-    #Given the new case details are stored in firestore
-    #And CC SVC fetches the case details from firestore
-    #When a fulfilment request, for a product code, is made for the case
-    #Then a fulfilment request event is sent to RM
-    #
